requirements.txt
django>=4.2,<4.3
djangorestframework>=3.14
channels>=4.0,<5
channels-redis>=4.3,<5
django-cors-headers>=4.0
asgiref>=3.9.1,<4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
django-environ>=0.9
uvicorn>=0.22.0
daphne>=4.0,<5
redis>=5.0
dj-database-url>=2.1
gunicorn>=21.0
whitenoise>=6.5
